Adding support for raw Bluray discs
ngl not going super well so far, this works with bdinfo in a docker container so that needs to be installed ahead of time.

TrueHD and audio codecs in general are being a total PITA so that will need to be rewritten later.

But in its current state, full bluray discs can be uploaded to all supported trackers